+++ apis/study.py
@@ -1,37 +1,17 @@
-from flask import request
+from flask import request, g
 from flask_restx import Namespace, Resource, fields
 
-from model import Study, db
+from model import Study, db, User, StudyContributor
+from .authentication import is_granted
 
-api = Namespace("study", description="study operations", path="/")
+api = Namespace("Study", description="Study operations", path="/")
 
-owner = api.model(
-    "Owner",
-    {
-        "id": fields.String(required=True),
-        "affiliations": fields.String(required=True),
-        "email": fields.String(required=True),
-        "first_name": fields.String(required=True),
-        "last_name": fields.String(required=True),
-        "orcid": fields.String(required=True),
-        "roles": fields.List(fields.String, required=True),
-        "permission": fields.String(required=True),
-        "status": fields.String(required=True),
-    },
-)
 
-study = api.model(
+study_model = api.model(
     "Study",
     {
-        "id": fields.String(required=True),
-        "name": fields.String(required=True),
-        "title": fields.String(required=True),
-        "description": fields.String(required=True),
-        "image": fields.String(required=True),
-        "size": fields.String(required=True),
-        "keywords": fields.String(required=True),
-        "last_updated": fields.String(required=True),
-        "owner": fields.Nested(owner, required=True),
+        "title": fields.String(required=True, default=""),
+        "image": fields.String(required=True, default=""),
     },
 )
 
@@ -41,77 +21,75 @@ class Studies(Resource):
     @api.doc("list_study")
     @api.response(200, "Success")
     @api.response(400, "Validation Error")
-    @api.param("id", "The study identifier")
-    # @api.marshal_list_with(study)
+    # @api.marshal_with(study_model)
     def get(self):
-        studies = Study.query.all()
+        """this code ensure each user access and see only allowed studies"""
+        # studies = Study.query.filter(
+        #     Study.study_contributors.any(User.id == g.user.id)
+        # ).all()
+        # studies = Study.query.filter(User.id == g.user.id).all()
+        study_contributors = StudyContributor.query.filter(
+            StudyContributor.user_id == g.user.id
+        ).all()  # Filter contributors where user_id matches the user's id
+        study_ids = [contributor.study_id for contributor in study_contributors]
+
+        studies = Study.query.filter(Study.id.in_(study_ids)).all()
         return [s.to_dict() for s in studies]
 
+    @api.expect(study_model)
+    @api.response(200, "Success")
+    @api.response(400, "Validation Error")
     def post(self):
         add_study = Study.from_data(request.json)
         db.session.add(add_study)
+        study_id = add_study.id
+        study_ = Study.query.get(study_id)
+        study_contributor = StudyContributor.from_data(study_, g.user, "owner")
+        db.session.add(study_contributor)
         db.session.commit()
-        return add_study.to_dict()
+        return study_.to_dict()
 
 
 @api.route("/study/<study_id>")
 class StudyResource(Resource):
-    @api.doc("update study")
+    @api.doc("get study")
     @api.response(200, "Success")
     @api.response(400, "Validation Error")
-    @api.param("id", "The study identifier")
     # @api.marshal_with(study)
     def get(self, study_id: int):
         study1 = Study.query.get(study_id)
         return study1.to_dict()
 
+    @api.expect(study_model)
+    @api.response(200, "Success")
+    @api.response(400, "Validation Error")
     def put(self, study_id: int):
         update_study = Study.query.get(study_id)
-        # if not addStudy.validate():
-        #     return 'error', 422
+        if not is_granted("update_study", update_study):
+            return "Access denied, you can not modify", 403
+
         update_study.update(request.json)
         db.session.commit()
         return update_study.to_dict()
 
+    @api.response(200, "Success")
+    @api.response(400, "Validation Error")
     def delete(self, study_id: int):
-        delete_study = Study.query.get(study_id)
-        for d in delete_study.dataset:
+        study = Study.query.get(study_id)
+        if not is_granted("delete_study", study):
+            return "Access denied, you can not delete study", 403
+        for d in study.dataset:
             for version in d.dataset_versions:
                 version.participants.clear()
-        for d in delete_study.dataset:
+        for d in study.dataset:
             for version in d.dataset_versions:
                 db.session.delete(version)
             db.session.delete(d)
-        for p in delete_study.participants:
+        for p in study.participants:
             db.session.delete(p)
-        db.session.delete(delete_study)
+        db.session.delete(study)
         db.session.commit()
-        return "", 204
-
-
-# @api.route("/viewProfile", methods=["GET"])
-# def viewProfile():
-#     dic = {
-#         "username": "admin",
-#         "email": "aydan.gasimova2@gmail.com",
-#         "fullname": "Aydan Gasimova",
-#         "image": f" https://api.dicebear.com/5.x/shapes/svg?seed=$"
-#         f"{str(random.randint(0,1000))}",
-#         "institution": "CALMI2",
-#         "location": "San Diego, CA",
-#         "password": "admin",
-#         "timezone": "(GMT-11:00) Midway Island",
-#     }
-#     return jsonify(dic)
-#
-
-
-# @study.route("/viewProfile", methods=["POST"])
-# def update_user_profile():
-#     data = request.json
-#
-#     if data is not None:
-#         data["id"] = 3
-#
-#     return jsonify(data), 201
-#
+        studies = Study.query.filter(
+            Study.study_contributors.any(User.id == g.user.id)
+        ).all()
+        return [s.to_dict() for s in studies], 201
